feat: get business type from SUNAT+RUC
