+++ backend/Controllers/Client/ClientModule.cs
@@ -1,0 +1,10 @@
+namespace PeruControl.Controllers;
+
+public class ClientModule : IModule
+{
+    public IServiceCollection SetupModule(IServiceCollection services, IConfiguration configuration)
+    {
+        services.AddScoped<ClientService>();
+        return services;
+    }
+}
